refactor(AppUnitTest): 출력 기능 단위 테스트 파일로 분리
